relocated the Trainer scripts to Training folder
The trainer scripts are for training various models in the system. They were previously located in the Marie/Utils folder together with the model class files to simplify folder structure in the HPC (So that one folder Marie/Utils contains everything needed).

Now for the clearance of the folder structure, the trainer scripts are relocated to the separated folder Training/Trainers