dev-forecasting-ntu-scenario-compatibility: Make forecastingagent compatible with the kg of ntu scenario
The agent throws errors when: (a) multiple measures are associated with the same time series and (b) the timeformat in the kg is not in iso format or python syntax. The timeformat used in the ntu scenario is now included in the dict and the query is changed so that the fc_iri has to be in a OM_HASVALUE relationship with the data_iri.